int bigint

+++ database/migrations/2019_12_03_135039_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table -> integer('jan') -> length(13) -> nullable();
+            $table -> bigInteger('jan') -> length(13) -> nullable();
             $table->string('pic') -> nullable();
             $table->text('comment') -> nullable();
             $table->integer('price');
